Chapter 04/03. Docker/exercise/scene_predictor/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
Pillow==9.2.0
requests==2.28.1
torch==1.12.0
torchvision==0.13.0
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
